fix: add STS regional Endpoint support with fallback path
As global endpoint is under the deprecation path, we will use regional STS endpoint
by default. To make sure our current customers will not be affected by this change,
whenever regional STS endpoint is not available, we will always fallback to the STS
endpoint of a particular partition which is always active.