fix: don't monitor files in file state folder
File state folder contains the state files generated by agent. These files
keep track of the reading offset for the log files tracked by agent. Users
are not supposed to monitor the files in the file state folder. With some
misconfiguration, if the state files are also monitored by agent, new state
files tracking existing state files will be created. This is not desired
behavior and should be not allowed.