chore: retrieve metric type outside prometheus
* Get metric type from metaDataCache of each scraper target
(similar to what open-telmetry-collector did:
https://github.com/open-telemetry/opentelemetry-collector/blob/master/receiver/prometheusreceiver/internal/metadata.go)

* The job name specified in prometheus config is used by scraper manager to
index the corresponding target group in metadata cache. However when the job is
replaced in relabel_config, we get the replaced job in metrics. The replaced job can't be
used to retrieve the metric metadata as the index is based on the old job name.
In this case, we go through all the target groups and checks the updated labels
to figure out the relevant target group.