prom: Fix test by adding saved instance label to mock
In metadataServiceImpl we were using `instance`, but like job it can be
changed in both `relabel_configs` and `metric_relabel_configs` so we can
only rely on our 'internal' label.

The labels saved in target are those after `relabel_config`, which is
how we were able to look up for relabeld job. Now we are using the saved
label so this loop and check all the targets in group can be avoid for
job. But the same logic is still required for instance, and since we are
passing instance value before relabel, we look up using or
savedScrapeInstanceLabel key instead of model.InstanceLabel