start: Skip pidfile when running in a container.
When the agent is running in a container, the use and management of a
pidfile shouldn't be necessary. By making this change, it moves the
agent container one step closer to being able to run entirely with a read
only root file system. While the pid file functionality doesn't prevent
running with a read only file system, it generates an unnecessary error
line in the logs:

2021-05-09T17:03:00Z E! Unable to create pidfile: open /opt/aws/amazon-cloudwatch-agent/var/amazon-cloudwatch-agent.pid: read-only file system

While it is possible to provide a small scratch filesystem for the pid
file, this seems entirely unnecessary.