chore: Enable build mode override via env variable
Go compiler has switched to internal linking mode when building in PIE
mode since version 1.15 (https://golang.org/doc/go1.15#linker). This had
a side effect of not able to compile a purely statically linked PIE
binary with CGO disabled, see: https://github.com/golang/go/issues/40711

As a result, the agent binaries build with PIE mode will fail to run in
a container that is built from SCRATCH. In this CR we allow the users to
override the build mode with environment variable `CWAGENT_BUILD_MODE`,
allowing a non-PIE binary to be built to be used in a container without
libc.
For example, a non-pie agent can be built with:
```
CWAGENT_BUILD_MODE=default make
```