Performance fix for interaction of Container Insights code with Kubernetes API server.
When running in a Kubernetes cluster, one leader-elected instance of the agent stays up to date with metadata such as
the full list of pods and nodes in the cluster, in order to post metrics for Container Insights.  It does this using a
Go client called the Reflector which performs API calls against the Kubernetes API server for each resource type.  The
Reflector works in two phases:
1. Perform an initial List call to receive the current set of resources (e.g., pods), at whatever version the API server
   happens to return.  The List call is paginated in case there are a large number of resources, so this phase might
   entail several API calls.
2. Thereafter, perform long-lived Watch calls to receive updates from the server whenever the set of resources changes,
   e.g. if a new pod is launched or one is deleted.

More information about the behavior of API calls here: https://kubernetes.io/docs/reference/using-api/api-concepts/

The K8S API server can serve List calls either from its cache or from etcd, its underlying data store, depending on the
consistency of results requested.  In the specific scenario where the request is paginated but resourceVersion is set to
0, the API server will serve from cache but ignore the pagination parameter, and thus return all resources in one huge
response.  For a customer with a large number of resources, the large response was causing the agent to exceed its
memory limit and restart, thus starting the List phase again on a new pod which would do the same expensive List API
call over again.  This would repeat until one pod managed to process the response.  In the meantime, the expensive calls
would drive up the CPU load on the API server nodes and starve out other requests.

This change forces the Reflector to set resourceVersion to empty instead of 0.  This causes the API server to send the
call to etcd instead of serving from cache, which respects pagination.  It does introduce more load on etcd, which the
Reflector was explicitly trying to avoid, but these calls should be infrequent - once the initial List call succeeds,
the Reflector will Watch until the pod it's running on is killed.