ECS Service Discovery: Fixup implicit network mode (#385)
I originally submitted https://github.com/aws/amazon-cloudwatch-agent/issues/330
with the goal of correcting the service discovery behavior of a task when its
definition did not specify a network mode.

Very graciously, @javabrett submitted a fix for this via
https://github.com/aws/amazon-cloudwatch-agent/pull/335. Upon testing this out
today, I found out we're not _quite_ there.

Crucially, when network mode is undefined in a task definition, it will parse as
`nil`, not the empty string. Correcting this in our unit tests reveals the
behavior isn't working as desired, since the `getPrivateIp` function isn't
properly handling the `nil` case. I believe this addresses the issue.