Create weekly workflow to sync latest upstream changes  (#33)
* Create weekly workflow to sync latest changes from upstream

1. get-latest: pushes latest upstream changes onto local branch "repo-sync"
2. create-pr: creates a pr from "repo-sync" to "main"

* Added a unique suffix to every branch and PR

With this setup, every workflow has a unique id, resulting in a separate branch and PR for every trigger. This avoids the edge case where someone is in the middle of resolving merge conflicts--and a new automated PR comes in and overwrites their unfinished work.

For the id, I used run_number.run_attempt because it's human readable and maintains uniqueness even when workflows are re-run

* Replaced all uses of PAT with GITHUB_TOKEN

So long as PR creation/deletion thru GitHub actions is enabled at the org level, then GITHUB_TOKEN can replace PAT in this context

* Created job to define pr branch name using last commit hash

Branch name must be defined inside a job and exported to jobs.output instead of an environmental variable, because that's the best practice to pass information between jobs.