In workflow step repo-sync/create-pr, upgraded token from GITHUB_TOKEN to PAT (#41)
Token must be upgraded because modifying workflow files requires the 'worfklow' scope, which is only enabled in PAT. Unfortunately, our org currently disables fine-grained PATs, which is the best practice because it limits the scopes to this repository. Instead, I am forced to use a general purpose PAT for now, which is tied to a user (me).