[E2E] Comment CW Agent change check
The image is changing but the old image value is incorrectly showing the same as the new image. This is likely an indication of some weird behaviour when overriding the CW Agent image that causes the new image to persist between separate installations. Will take a follow up item to investigate, but making this change to unblock the release.