Adhere to SOURCE_DATE_EPOCH specification to fix reproducibility
Reproducible builds are ones where the same source code input produces the same
binary output. This helps with caching, authentication of binaries, etc.
SOURCE_DATE_EPOCH is a standardised variable which allows the build setup to
specify date and time to be used in place of "current time" such that it will
not change unless the source code changes.

See:
https://reproducible-builds.org/specs/source-date-epoch/
https://reproducible-builds.org/docs/source-date-epoch/

Signed-off-by: Oleksandr Hnatiuk <ohnatiuk@cisco.com>