[ApplicationSignals] Support copying resource attributes to metric attributes for EMF export
In this change, we are supporting new attribute copying behaviour for ApplicationSignals. This will allow customers to specify which/all attributes they want to be present in their EMF records.

If the following key is present in the resource attributes: aws.application_signals.metric_resource_keys, we will copy attributes specified in the value (or all attributes if all_attributes) from resource to metric. Note that we do not copy attributes that are already copied/normalized by exisiting ApplicationSignals processor logic, to avoid data duplication. Also note that we prefix all copied attribute keys with otel.resource to avoid collisions with existing metric attributes.